+++ tests/unit/io/WKTReaderTest.cpp
@@ -317,4 +317,15 @@ void object::test<13>
 
 
 
+// EMPTY token with some white space
+template<>
+template<>
+void object::test<14>
+()
+{
+    GeomPtr geom(wktreader.read("MULTIPOINT( EMPTY, (10 10), (20 20))"));
+
+    ensure_equals(geom->getNumGeometries(), 3u);
+}
+
 } // namespace tut